requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy
psycopg2-binary
redis
python-dotenv
streamlit
langchain>=0.3.0
langchain-openai
langchain-community
openai
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-multipart
alembic
faiss-cpu
requests
pandas
plotly
reportlab
weasyprint
sendgrid
twilio
google-cloud-tasks
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0
locust>=2.17.0

